Update localization to Thai
This commit updates the localization of the application to Thai. The fallback language is set to English and the second language is set to Thai. All the French localization files have been deleted and the corresponding Thai localization files have been added. The middleware has been updated to redirect /th/home to /th.